Added support for stacking of directors
Directors can now use directors as backends. The actual backend and
director code was already capable of doing it, so I only had to
update the VCC to grok it. In doing so the hosts VTAILQ could be
dropped, and a new VTAILQ was added to be able to do director
destruction in opposite order from initialization.