Drop old grace objects when we have a new object that matches the Vary
Also add EXP_Remove, to allow removing objects with touching them.

This patch makes it possible to have a very short TTL, but a long grace.
Normally, if you were to set a 1s TTL and a 86400s grace, you would end
up with 86400 objects on a single objecthead, not even taking into
account Vary. With this patch, any object found in grace is kept with
the request, and upon succesful retrieval of a new object from the
backend, the grace object is removed.

EXP_Remove was added so that the TTL and grace on the object aren't
touched. This, in combination with some other work we've done, prevents
pages from getting dirtied and thus having to be written out to disk
in a environment where the cache cannot be purely kept in memory.

VRY_Compare compares 2 vary strings for equality (taking gzip into
account) and was added so that the grace object is only dropped if its
vary matches to the vary on the new object.